% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCollections.R
\name{laads_collections}
\alias{laads_collections}
\title{Available collections.}
\usage{
laads_collections()
}
\value{
A data.frame (tibble) with the Name and Description of available collections
}
\description{
Available collections.
}
\examples{
\dontrun{
laads_collections()
}
}

