% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateCoverage.R
\name{laads_data_coverage}
\alias{laads_data_coverage}
\title{Range of dates covered by the given product in the given collection.}
\usage{
laads_data_coverage(collection = "5", product = "MCD15A2")
}
\arguments{
\item{collection}{character, collection name as given by laads_collections()}

\item{product}{character, product name as given by laads_products()}
}
\value{
A data.frame (tibble) with POSIXct start and end of the data collection.
}
\description{
Range of dates covered by the given product in the given collection.
}
\details{
The range does not account for gaps in the data collection.
}
\examples{
\dontrun{
laads_data_coverage(collection = "5", product = "MCD15A2")
}
}

