% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileProperties.R
\name{laads_file_properties}
\alias{laads_file_properties}
\title{Title}
\usage{
laads_file_properties(file_ids = "299343600")
}
\arguments{
\item{file_ids}{file IDs for the data files, either a single one or a vector of IDs. For getting IDs see \code{laads_search_files}.}
}
\value{
A data.frame (tibble) containing data file properties, including name, id, size, format, ingest time, data start time, and online status.
}
\description{
Title
}
\examples{
\dontrun{
files <- laads_search_files(product = "MCD15A2",
                              start_time = "2010-01-01",
                              end_time = "2010-01-10",
                              coords_or_tiles = "global",
                              day_night_both = "DNB")$file_id
laads_file_properties(file_ids = files[1:10])}
}

