% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxSearchResults.R
\name{laads_max_search_results}
\alias{laads_max_search_results}
\title{Maximum number of file IDs that will be returned from a search.}
\usage{
laads_max_search_results()
}
\value{
A numeric
}
\description{
Maximum number of file IDs that will be returned from a search.
}
\examples{
\dontrun{
laads_max_search_results()
}
}

