% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProducts.R
\name{laads_products}
\alias{laads_products}
\title{Available products.}
\usage{
laads_products()
}
\value{
A data.frame (tibble) with the Name,
Description and Default Collections of available products.
}
\description{
Available products.
}
\examples{
\dontrun{
laads_products()
}
}
