% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProductsByInstrument.R
\name{laads_products_instrument}
\alias{laads_products_instrument}
\title{Available product groups for a given instrument}
\usage{
laads_products_instrument(instrument = "AM1M", group = NULL)
}
\arguments{
\item{instrument}{character, the instrument short name as given by \code{laads_satellite_instruments()}.}

\item{group}{character, the group short name as given by \code{laads_product_groups()}.}
}
\value{
A data.frame (tibble) with the Name and Description of available products
}
\description{
Available product groups for a given instrument
}
\examples{
\dontrun{
laads_products_instrument(instrument = "AM1M", group = "LL3L4_A")
}
}

