% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSatelliteInstruments.R
\name{laads_satellite_instruments}
\alias{laads_satellite_instruments}
\title{Available satellite instruments.}
\usage{
laads_satellite_instruments()
}
\value{
A data.frame (tibble) with the Name and Description of available satellites/instruments
}
\description{
Available satellite instruments.
}
\examples{
\dontrun{
laads_satellite_instruments()
}
}
