% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchDatasets.R
\name{laads_search_datasets}
\alias{laads_search_datasets}
\title{Returns a list of products matching the supplied keywords.}
\usage{
laads_search_datasets(keywords = c("AOD", "MODIS"))
}
\arguments{
\item{keywords}{A vector of keywords (character).}
}
\value{
A data.frame (tibble) with time of update, author name and email,
 title of the product (which helps knowing what's inside),
start and end time of the product, and the product name.
}
\description{
Returns a list of products matching the supplied keywords.
}
\details{
The keywords are OR'd, and will match product shortname, longname, and name of any SDS's contained in the product.
}
\examples{
\dontrun{
laads_search_datasets(keywords = c("AOD", "MODIS"))
}
}
