% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchForFiles.R
\name{laads_search_files}
\alias{laads_search_files}
\title{Returns a list of file IDs matching the given search criteria.}
\usage{
laads_search_files(product = "MCD15A2", collection = NULL,
  start_time = "2010-01-01", end_time = "2015-05-02", north = NULL,
  south = NULL, east = NULL, west = NULL, coords_or_tiles = "global",
  day_night_both = "DNB")
}
\arguments{
\item{product}{product name(s), e.g. \code{"MCD43C1"} or \code{c("MCD43C1", "MCD43B4")}. See \code{laads_products}.}

\item{collection}{collection number. Default is the default collection for the product. See \code{laads_collection}.}

\item{start_time}{start date (and optional time) of the temporal window, "YYYY-MM-DD" or "YYYY-MM-DD hh:mm:ss".}

\item{end_time}{end date (and optional time) of the temporal window, "YYYY-MM-DD" or "YYYY-MM-DD hh:mm:ss".}

\item{north}{North boundary of the spatial window, -90 to 90 for coords or 0 to 17 for tiles.}

\item{south}{South boundary of the spatial window, -90 to 90 for coords or 0 to 17 for tiles.}

\item{east}{East boundary of the spatial window, -180 to 180 for coords or 0 to 35 for tiles.}

\item{west}{West boundary of the spatial window, -180 to 180 for coords or 0 to 35 for tiles.}

\item{coords_or_tiles}{"coords", "tiles" or "global", specifies whether the north/south/west/east parameters are coords or tiles. The north/south/west/east parameters are ignored if global.}

\item{day_night_both}{specifies whether to include data files which only have day data (D), only have night data (N), or only have both day and night data (B). Possible values are these combinations: "D", "N", "B", "DN", "DB", "NB", "DNB". Default is "DNB".}
}
\value{
A data.frame (tibble) with only one column, "file_id".
}
\description{
Returns a list of file IDs matching the given search criteria.
}
\examples{
\dontrun{
laads_search_files(product = "MCD15A2",
                     start_time = "2010-01-01",
                    end_time = "2015-05-02",
                    coords_or_tiles = "global",
                    day_night_both = "DNB")}

}

