% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollections.R
\name{laads_collections}
\alias{laads_collections}
\title{Available collections.}
\usage{
laads_collections(product = "MCD15A2")
}
\arguments{
\item{product}{character, the product short name as given by \code{laads_products()}.}
}
\value{
A data.frame (tibble) with the Name and Description of available collections
}
\description{
Available collections.
}
\examples{
\dontrun{
laads_collections()
}
}

