% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProductGroups.R
\name{laads_product_groups}
\alias{laads_product_groups}
\title{Available products for a given instrument}
\usage{
laads_product_groups(instrument = "AM1M")
}
\arguments{
\item{instrument}{character, the product short name as given by \code{laads_satellite_instruments()}.}
}
\value{
A data.frame (tibble) with the Name and Description of available products
}
\description{
Available products for a given instrument
}
\examples{
\dontrun{
laads_collections()
}
}

